refactor: move donate history list to donate tab
